requirements.txt
streamlit
streamlit_chat
openai
langchain
azure-search-documents==11.4.0b6
azure-search-documents
azure-identity
python-dotenv
tiktoken
azure-storage-blob
langchain
unstructured
unstructured[pdf]